requirements.txt
boolean.py==4.0
CacheControl==0.14.1
certifi==2024.8.30
charset-normalizer==3.4.0
cyclonedx-python-lib==7.6.2
defusedxml==0.7.1
filelock==3.16.1
html5lib==1.1
idna==3.10
instaloader==4.14
license-expression==30.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
packageurl-python==0.16.0
packaging==24.2
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
py-serializable==1.1.2
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sortedcontainers==2.4.0
toml==0.10.2
urllib3==2.2.3
webencodings==0.5.1
